Refactored include strategy.
