+++ Client/CMakeLists.txt
@@ -5,6 +5,7 @@ project(R_Type_Client)
 add_subdirectory(UnitTests)
 add_subdirectory(SFML EXCLUDE_FROM_ALL)
 include_directories(SYSTEM "SFML/include/")
+include_directories(../Shared/LibraryLoader/ ../Shared/Libs/Interfaces ../Shared/Include)
 
 
 IF (MSVC)
