+++ Server/src/main.cpp
@@ -2,7 +2,6 @@
 // Created by hippolyteb on 11/20/16.
 //
 
-
 int main()
 {
     return (0);
