+++ Shared/Include/ITweeningCurve.hpp
