+++ Shared/Time/TimeRef.cpp
@@ -3,7 +3,7 @@
 //
 
 #include <chrono>
-#include "TimeRef.hpp"
+#include "../Include/TimeRef.hpp"
 
 TimeRef::TimeRef() : _current(std::chrono::milliseconds()) { }
