+++ Shared/Tween/Curve/EaseInOutCurve.cpp
@@ -3,7 +3,7 @@
 //
 
 #include <math.h>
-#include "EaseInOutCurve.hpp"
+#include "../../Include/EaseInOutCurve.hpp"
 
 double EaseInOutCurve::Curve(const double &d) const {
     return -0.5 * (cos(M_PI*d) - 1);
