+++ Shared/Tween/Curve/LinearCurve.cpp
@@ -2,7 +2,7 @@
 // Created by hippolyteb on 11/24/16.
 //
 
-#include "LinearCurve.hpp"
+#include "../../Include/LinearCurve.hpp"
 
 double LinearCurve::Curve(double const &d) const {
     return d;
