+++ Shared/Tween/Curve/SmoothStepCurve.cpp
@@ -2,7 +2,7 @@
 // Created by hippolyteb on 11/24/16.
 //
 
-#include "SmoothStepCurve.hpp"
+#include "../../Include/SmoothStepCurve.hpp"
 
 double SmoothStepCurve::Curve(const double &d) const {
     return (d*d *(3 - 2*d));
