+++ Shared/UnitTests/Time/Timer.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 #include <thread>
-#include "../../Time/Timer.hpp"
+#include "../../Include/Timer.hpp"
 
 TEST(Tests_Time, CreateTimerFromCurrentTime) {
     auto thread = new std::thread([&] {
